+++ noodles-fastq/src/record/definition.rs
@@ -1,18 +1,20 @@
+use bstr::{BStr, BString};
+
 /// A FASTQ record definition.
 ///
 /// A definition represents a definition line, i.e., a read name and, optionally, a description.
 #[derive(Clone, Debug, Default, Eq, PartialEq)]
 pub struct Definition {
-    name: Vec<u8>,
-    description: Vec<u8>,
+    name: BString,
+    description: BString,
 }
 
 impl Definition {
     /// Creates a FASTQ record definition.
     pub fn new<N, D>(name: N, description: D) -> Self
     where
-        N: Into<Vec<u8>>,
-        D: Into<Vec<u8>>,
+        N: Into<BString>,
+        D: Into<BString>,
     {
         Self {
             name: name.into(),
@@ -21,22 +23,22 @@ impl Definition {
     }
 
     /// Returns the record name.
-    pub fn name(&self) -> &[u8] {
-        &self.name
+    pub fn name(&self) -> &BStr {
+        self.name.as_ref()
     }
 
     /// Returns a mutable reference to the record name.
-    pub fn name_mut(&mut self) -> &mut Vec<u8> {
+    pub fn name_mut(&mut self) -> &mut BString {
         &mut self.name
     }
 
     /// Returns the description.
-    pub fn description(&self) -> &[u8] {
-        &self.description
+    pub fn description(&self) -> &BStr {
+        self.description.as_ref()
     }
 
     /// Returns a mutable reference to the description.
-    pub fn description_mut(&mut self) -> &mut Vec<u8> {
+    pub fn description_mut(&mut self) -> &mut BString {
         &mut self.description
     }