+++ noodles-fastq/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * fastq/record/definition: Change underlying types to byte strings
+    (`BString`).
+
 ## 0.13.0 - 2024-07-14
 
 ### Removed